backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.8.3
anthropic==0.40.0
httpx==0.27.2
Pillow==10.4.0
firebase-admin==6.5.0
twilio==9.3.5
